isolate StorageHandle from main.rs & add small tests
