+++ Cargo.toml
@@ -16,3 +16,6 @@ structopt = "0.2.11"
 trackable = "0.2.20"
 rustyline = "2.1.0"
 regex = "1.0.5"
+
+[dev-dependencies]
+tempdir = "*"
